// lib/api/__generated__/models/GetProfileByUsername200Response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * RealWorld Conduit API
 * Conduit API documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Profile } from './Profile';
import {
    ProfileFromJSON,
    ProfileFromJSONTyped,
    ProfileToJSON,
} from './Profile';

/**
 * 
 * @export
 * @interface GetProfileByUsername200Response
 */
export interface GetProfileByUsername200Response {
    /**
     * 
     * @type {Profile}
     * @memberof GetProfileByUsername200Response
     */
    profile: Profile;
}

/**
 * Check if a given object implements the GetProfileByUsername200Response interface.
 */
export function instanceOfGetProfileByUsername200Response(value: object): boolean {
    if (!('profile' in value)) return false;
    return true;
}

export function GetProfileByUsername200ResponseFromJSON(json: any): GetProfileByUsername200Response {
    return GetProfileByUsername200ResponseFromJSONTyped(json, false);
}

export function GetProfileByUsername200ResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetProfileByUsername200Response {
    if (json == null) {
        return json;
    }
    return {
        
        'profile': ProfileFromJSON(json['profile']),
    };
}

export function GetProfileByUsername200ResponseToJSON(value?: GetProfileByUsername200Response | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'profile': ProfileToJSON(value['profile']),
    };
}