// lib/api/__generated__/models/UpdateCurrentUserRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * RealWorld Conduit API
 * Conduit API documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { UpdateUser } from './UpdateUser';
import {
    UpdateUserFromJSON,
    UpdateUserFromJSONTyped,
    UpdateUserToJSON,
} from './UpdateUser';

/**
 * 
 * @export
 * @interface UpdateCurrentUserRequest
 */
export interface UpdateCurrentUserRequest {
    /**
     * 
     * @type {UpdateUser}
     * @memberof UpdateCurrentUserRequest
     */
    user: UpdateUser;
}

/**
 * Check if a given object implements the UpdateCurrentUserRequest interface.
 */
export function instanceOfUpdateCurrentUserRequest(value: object): boolean {
    if (!('user' in value)) return false;
    return true;
}

export function UpdateCurrentUserRequestFromJSON(json: any): UpdateCurrentUserRequest {
    return UpdateCurrentUserRequestFromJSONTyped(json, false);
}

export function UpdateCurrentUserRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateCurrentUserRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'user': UpdateUserFromJSON(json['user']),
    };
}

export function UpdateCurrentUserRequestToJSON(value?: UpdateCurrentUserRequest | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'user': UpdateUserToJSON(value['user']),
    };
}